chore(pages): reorganize Professor into academic_management/exams/dashboard (QA) and update routes/imports
