+++ client/src/routes/routes.tsx
@@ -8,10 +8,10 @@ import Interview from "../pages/student/reinforcement/interview";
 import { Reinforcement } from "../pages/student/reinforcement/reinforcement";
 import Test from "../pages/student/reinforcement/test";
 import PeriodDetailPage from "../pages/periods-classes/PeriodDetailPage";
-import CoursePeriodsPage from "../pages/courses/CoursePeriodsPage";
+import CoursePeriodsPage from "../pages/professor/academic_management/CoursePeriodsPage";
 import DashboardPage from "../pages/common/DashboardPage";
-import ExamsCreatePage from "../pages/exams/ExamCreatePage";
-import ExamManagementPage from "../pages/exams/ExamManagementPage";
+import ExamsCreatePage from "../pages/professor/exams/ExamCreatePage";
+import ExamManagementPage from "../pages/professor/exams/ExamManagementPage";
 import ForgotPasswordPage from "../pages/common/ForgotPassword";
 import Login from "../pages/common/Login";
 import SettingsPage from "../pages/common/SettingsPage";
